Before viewer analizing...

+++ lsd_slam_viewer/src/PointCloudViewer.cpp
@@ -56,8 +56,8 @@ PointCloudViewer::PointCloudViewer()
         setPathKey((int)Qt::Key_9,(int)9);
 
 
-	currentCamDisplay = 0;
-	graphDisplay = 0;
+    currentCamDisplay   = 0;
+    graphDisplay        = 0;
 
 
 	for(int i=0;i<10;i++)
@@ -327,7 +327,8 @@ void PointCloudViewer::keyPressEvent(QKeyEvent *e)
     	  meddleMutex.lock();
 
 
-          qreal x,y,z;
+  //      qreal x,y,z;
+          float x,y,z;
     	  camera()->frame()->getPosition(x,y,z);
     	  animationList.push_back(AnimationObject(false, lastAnimTime, 2, qglviewer::Frame(qglviewer::Vec(0,0,0), camera()->frame()->orientation())));
     	  animationList.back().frame.setPosition(x,y,z);
